requirements.txt
asttokens==2.4.1
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
colorama==0.4.6
comm==0.2.2
coverage==7.4.4
cryptography==42.0.5
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
executing==2.0.1
filelock==3.13.3
greenlet==3.0.3
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.3
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==25.1.2
sentry-sdk==1.43.0
six==1.16.0
SQLAlchemy==2.0.28
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
